+++ askbot/conf/moderation.py
@@ -4,6 +4,7 @@ from askbot.conf.settings_wrapper import settings
 from askbot.conf.super_groups import DATA_AND_FORMATTING
 from askbot.deps.livesettings import ConfigurationGroup
 from askbot.deps.livesettings import BooleanValue
+from askbot.deps.livesettings import LongStringValue
 from askbot.deps.livesettings import StringValue
 from django.core.cache import cache
 from django.utils.translation import ugettext_lazy as _
@@ -28,6 +29,19 @@ settings.register(
         default='flags',
         description=_('Content moderation method'),
         help_text=_("Audit is made after the posts are published, pre-moderation prevents publishing before moderator's decision.")
+    )
+)
+
+settings.register(
+    LongStringValue(
+        MODERATION,
+        'FORBIDDEN_PHRASES',
+        default='',
+        description=_('Reject all posts with these phrases'),
+        help_text=_('Enter one phrase per line (case-insensitive). '
+            'Posts with these phrases will be rejected '
+            'without moderation.'
+        )
     )
 )